+++ engine/src/engine.rs
@@ -3,5 +3,5 @@ pub trait GameEngine {
     type State;
 
     fn take_action(&self, game_state: &Self::State, action: &Self::Action) -> Self::State;
-    fn is_terminal_state(&self, game_state: &Self::State) -> Option<f64>;
+    fn is_terminal_state(&self, game_state: &Self::State) -> Option<f32>;
 }
