+++ model/src/position_metrics.rs
@@ -3,6 +3,6 @@ use super::node_metrics::NodeMetrics;
 #[derive(Debug)]
 pub struct PositionMetrics<S, A> {
     pub game_state: S,
-    pub score: f64,
+    pub score: f32,
     pub policy: NodeMetrics<A>
 }
