+++ model/src/tensorflow_serving/model.rs
@@ -36,9 +36,9 @@ pub struct TensorflowServingModel<S,A,E,Map>
 }
 
 pub trait Mapper<S,A> {
-    fn game_state_to_input(&self, game_state: &S) -> Vec<Vec<Vec<f64>>>;
-    fn policy_metrics_to_expected_input(&self, policy: &NodeMetrics<A>) -> Vec<f64>;
-    fn policy_to_valid_actions(&self, game_state: &S, policy_scores: &Vec<f64>) -> Vec<ActionWithPolicy<A>>;
+    fn game_state_to_input(&self, game_state: &S) -> Vec<Vec<Vec<f32>>>;
+    fn policy_metrics_to_expected_input(&self, policy: &NodeMetrics<A>) -> Vec<f32>;
+    fn policy_to_valid_actions(&self, game_state: &S, policy_scores: &Vec<f32>) -> Vec<ActionWithPolicy<A>>;
 }
 
 impl<S,A,E,Map> TensorflowServingModel<S,A,E,Map>
@@ -74,7 +74,7 @@ where
                     if i == 0 && elapsed_mills >= 5_000 {
                         let num_nodes = batching_model_ref.take_num_nodes_analysed();
                         let (min_batch_size, max_batch_size) = batching_model_ref.take_min_max_batch_size();
-                        let nps = num_nodes as f64 * 1000.0 / elapsed_mills as f64;
+                        let nps = num_nodes as f32 * 1000.0 / elapsed_mills as f32;
                         let now = Utc::now().format("%H:%M:%S").to_string();
                         println!(
                             "TIME: {}, NPS: {:.2}, Min Batch Size: {}, Max Batch Size: {}",
@@ -549,12 +549,12 @@ where
 
 #[derive(Serialize)]
 struct RequestImage {
-    input_image: Vec<Vec<Vec<f64>>>
+    input_image: Vec<Vec<Vec<f32>>>
 }
 
 #[derive(Debug, Deserialize)]
 struct PredictionResults {
-    predictions: Vec<HashMap<String,Vec<f64>>>
+    predictions: Vec<HashMap<String,Vec<f32>>>
 }
 
 fn get_model_url(model_info: &ModelInfo) -> String {
