+++ ponder/src/ponder.rs
@@ -16,9 +16,9 @@ use model::model::Model;
 #[derive(Debug)]
 pub struct PonderOptions {
     pub visits: usize,
-    pub cpuct_base: f64,
-    pub cpuct_init: f64,
-    pub cpuct_root_scaling: f64
+    pub cpuct_base: f32,
+    pub cpuct_init: f32,
+    pub cpuct_root_scaling: f32
 }
 
 pub struct Ponder {}
@@ -54,7 +54,7 @@ impl Ponder
                 None,
                 0.0,
                 1.0,
-                |_,_,_,Nsb,is_root| (((Nsb as f64 + cpuct_base + 1.0) / cpuct_base).ln() + cpuct_init) * if is_root { cpuct_root_scaling } else { 1.0 },
+                |_,_,_,Nsb,is_root| (((Nsb as f32 + cpuct_base + 1.0) / cpuct_base).ln() + cpuct_init) * if is_root { cpuct_root_scaling } else { 1.0 },
                 |_,_| 0.0,
                 rng::create_rng_from_uuid(uuid),
             )
